requirements.txt
Flask>=2.2.5
Flask-Login>=0.6.2
cryptography>=41.0.2
google-auth>=2.20.0
gspread>=5.10.0
